refactor(models integer field): apply decimal field to models instead of integer field

+++ glucose_app/injection/models.py
@@ -2,6 +2,6 @@ from django.db import models
 
 class Injection(models.Model):
 
-    ui = models.IntegerField()
+    ui = models.DecimalField(max_digits=4, decimal_places=2)
     timestamp = models.DateTimeField()
     pacient = models.ForeignKey("pacient.Pacient", null=True, on_delete=models.CASCADE)
